+++ Source/MQTTnet.Benchmarks/SerializerBenchmark.cs
@@ -39,7 +39,7 @@ namespace MQTTnet.Benchmarks
 
             _bufferWriter = new MqttBufferWriter(4096, 65535);
             _serializer = new MqttV3PacketFormatter(_bufferWriter, MqttProtocolVersion.V311);
-            _serializedPacket = _serializer.Encode(_packet).ToArray();
+            _serializedPacket = _serializer.Encode(_packet).Join();
         }
 
         [Benchmark]
